Introduce class hierarchy in BigQuery CaseInsensitiveMapping tests
